add yaml for everything
