+++ packages/interface/package.json
@@ -80,6 +80,7 @@
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3",
     "uuid": "^8.3.1",
+    "voca": "^1.4.0",
     "yup": "^0.32.8"
   },
   "optionalDependencies": {
@@ -97,6 +98,7 @@
     "e2e-test": "start-server-and-test dev http://127.0.0.1:8000 'cypress open'"
   },
   "devDependencies": {
+    "@types/voca": "^1.4.0",
     "babel-plugin-add-react-displayname": "^0.0.5",
     "babel-preset-gatsby": "^0.8.0",
     "gatsby-cypress": "^0.8.0"
